v1/grpc_server/requirements.txt
grpc-cpp
grpcio-tools
numpy
pandas
keras
tensorflow

